Fixing the observations of extraChallenge
