+++ src/main/java/extraChallenge/Location.java
@@ -1,0 +1,31 @@
+package extraChallenge;
+
+public class Location {
+
+    private double lon;
+    private double lat;
+
+    public Location() {
+    }
+
+    public Location(double lon, double lat) {
+        this.lon = lon;
+        this.lat = lat;
+    }
+
+    public double getLon() {
+        return lon;
+    }
+
+    public void setLon(double lon) {
+        this.lon = lon;
+    }
+
+    public double getLat() {
+        return lat;
+    }
+
+    public void setLat(double lat) {
+        this.lat = lat;
+    }
+}
